Revamp UI examples with improved styling and structure
Updated all example components to use enhanced layouts, consistent background colors, and improved widget structure. Added descriptive headers and instructions to each example. Refactored popover, menu, select, and dialog examples for better usability and visual feedback, including new animation and focus/hover effects. Removed unused dialog result display and modernized select examples with themed variants.
