+++ example/lib/api/naked_popover.0.dart
@@ -1,121 +1,147 @@
 import 'package:flutter/material.dart';
 import 'package:naked_ui/naked_ui.dart';
 
-class BasicPopoverExample extends StatelessWidget {
-  const BasicPopoverExample({super.key});
+void main() {
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Basic Popover Example'),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const Text(
-              'Click the button below to show a popover',
-              style: TextStyle(fontSize: 16),
-            ),
-            const SizedBox(height: 24),
-            NakedPopover(
-              positioning: const OverlayPositionConfig(
-                alignment: Alignment.bottomCenter,
-                fallbackAlignment: Alignment.topCenter,
-              ),
-              popoverBuilder: (context, info) => Container(
-                constraints: const BoxConstraints(maxWidth: 280),
-                padding: const EdgeInsets.all(16),
-                decoration: BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.circular(8),
-                  border: Border.all(color: Colors.grey.shade300),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.black.withValues(alpha: 0.1),
-                      blurRadius: 8,
-                      offset: const Offset(0, 4),
-                    ),
-                  ],
-                ),
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    const Text(
-                      'Popover Content',
-                      style: TextStyle(
-                        fontSize: 16,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    const Text(
-                      'This is the content inside the popover. You can put any widget here.',
-                      textAlign: TextAlign.center,
-                    ),
-                    const SizedBox(height: 12),
-                    ElevatedButton(
-                      onPressed: () => Navigator.of(context).pop(),
-                      child: const Text('Close'),
-                    ),
-                  ],
+    return MaterialApp(
+      home: Scaffold(
+        backgroundColor: Colors.grey.shade50,
+        body: const Center(
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text(
+                'Simple Popover',
+                style: TextStyle(
+                  fontSize: 24,
+                  fontWeight: FontWeight.bold,
                 ),
               ),
-              child: Container(
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 20, vertical: 12),
-                decoration: BoxDecoration(
-                  color: const Color(0xFF3D3D3D),
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                child: const Text(
-                  'Show Popover',
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontSize: 16,
-                  ),
-                ),
+              SizedBox(height: 8),
+              Text(
+                'Click the button to open the popover',
+                style: TextStyle(color: Colors.grey),
               ),
-            ),
-            const SizedBox(height: 40),
-            Container(
-              padding: const EdgeInsets.all(16),
-              margin: const EdgeInsets.symmetric(horizontal: 20),
-              decoration: BoxDecoration(
-                color: Colors.blue.shade50,
-                borderRadius: BorderRadius.circular(8),
-                border: Border.all(color: Colors.blue.shade200),
-              ),
-              child: Column(
-                children: [
-                  Icon(
-                    Icons.info_outline,
-                    color: Colors.blue.shade600,
-                    size: 24,
-                  ),
-                  const SizedBox(height: 8),
-                  const Text(
-                    'Try it out!',
-                    style: TextStyle(
-                      fontSize: 16,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                  const SizedBox(height: 4),
-                  Text(
-                    'Click the button above to see the popover in action. You can dismiss it by tapping outside or using the close button.',
-                    textAlign: TextAlign.center,
-                    style: TextStyle(
-                      color: Colors.blue.shade700,
-                      fontSize: 14,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
+              SizedBox(height: 24),
+              PopoverExample(),
+            ],
+          ),
         ),
+      ),
+    );
+  }
+}
+
+class PopoverExample extends StatefulWidget {
+  const PopoverExample({super.key});
+
+  @override
+  State<PopoverExample> createState() => _PopoverExampleState();
+}
+
+class _PopoverExampleState extends State<PopoverExample> {
+  final _controller = MenuController();
+
+  @override
+  Widget build(BuildContext context) {
+    return NakedPopover(
+      controller: _controller,
+      consumeOutsideTaps: false,
+      positioning: const OverlayPositionConfig(
+        alignment: Alignment.bottomCenter,
+        fallbackAlignment: Alignment.topCenter,
+        offset: Offset(0, 8),
+      ),
+      popoverBuilder: (context, info) => const PopoverMenu(),
+      child: Container(
+        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+        decoration: BoxDecoration(
+          color: const Color(0xFF3D3D3D),
+          borderRadius: BorderRadius.circular(8),
+        ),
+        child: const Text(
+          'Show Popover',
+          style: TextStyle(
+            color: Colors.white,
+            fontSize: 14,
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class PopoverMenu extends StatelessWidget {
+  const PopoverMenu({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      constraints: const BoxConstraints(maxWidth: 280),
+      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+      decoration: BoxDecoration(
+        color: Colors.grey.shade50,
+        borderRadius: BorderRadius.circular(8),
+        border: Border.all(color: Colors.grey.shade300),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withValues(alpha: 0.05),
+            blurRadius: 8,
+            offset: const Offset(0, 4),
+          ),
+        ],
+      ),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Text(
+                'Popover Content',
+                style: TextStyle(
+                  fontSize: 16,
+                  fontWeight: FontWeight.w500,
+                ),
+              ),
+              Icon(Icons.close, size: 16),
+            ],
+          ),
+          const SizedBox(height: 4),
+          Text(
+            'This is the content inside the popover. You can put any widget here.',
+            textAlign: TextAlign.left,
+            style: TextStyle(
+              color: Colors.grey.shade600,
+            ),
+          ),
+          const SizedBox(height: 4),
+          Align(
+            alignment: Alignment.centerRight,
+            child: Container(
+              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+              decoration: BoxDecoration(
+                color: Colors.grey.shade200,
+                borderRadius: BorderRadius.circular(8),
+              ),
+              child: Text(
+                'Next',
+                style: TextStyle(
+                  color: Colors.grey.shade600,
+                ),
+              ),
+            ),
+          ),
+        ],
       ),
     );
   }
