+++ example/lib/api/naked_select.1.dart
@@ -23,32 +23,35 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       debugShowCheckedModeBanner: false,
       home: Scaffold(
-        backgroundColor: Colors.grey.shade50,
+        backgroundColor: Color(0xFF0D0D0D),
         body: Center(
-          child: Container(
-            padding: const EdgeInsets.all(24),
-            child: const Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Text(
-                  'Select with Checkmarks',
-                  style: TextStyle(
-                    fontSize: 24,
-                    fontWeight: FontWeight.bold,
-                  ),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text(
+                'CYBER SELECT',
+                style: TextStyle(
+                  fontSize: 24,
+                  fontWeight: FontWeight.bold,
+                  color: Color(0xFF00FF41),
+                  fontFamily: 'monospace',
+                  letterSpacing: 2,
                 ),
-                SizedBox(height: 8),
-                Text(
-                  'Options show checkmarks when selected',
-                  style: TextStyle(color: Colors.grey),
+              ),
+              SizedBox(height: 8),
+              Text(
+                '> INITIALIZE DATA STREAM_',
+                style: TextStyle(
+                  color: Color(0xFF00AA33),
+                  fontFamily: 'monospace',
                 ),
-                SizedBox(height: 32),
-                CheckmarkSelectExample(),
-              ],
-            ),
+              ),
+              SizedBox(height: 24),
+              CyberpunkSelectExample(),
+            ],
           ),
         ),
       ),
@@ -56,91 +59,113 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class CheckmarkSelectExample extends StatefulWidget {
-  const CheckmarkSelectExample({super.key});
+class CyberpunkSelectExample extends StatefulWidget {
+  const CyberpunkSelectExample({super.key});
 
   @override
-  State<CheckmarkSelectExample> createState() => _CheckmarkSelectExampleState();
+  State<CyberpunkSelectExample> createState() => _CyberpunkSelectExampleState();
 }
 
-class _CheckmarkSelectExampleState extends State<CheckmarkSelectExample> {
+class _CyberpunkSelectExampleState extends State<CyberpunkSelectExample> {
   String? _selectedValue;
 
-  // Available fruits
+  // Available cyber fruits
   static const fruits = [
-    Fruit(value: 'apple', label: 'Apple', emoji: '🍎'),
-    Fruit(value: 'banana', label: 'Banana', emoji: '🍌'),
-    Fruit(value: 'orange', label: 'Orange', emoji: '🍊'),
-    Fruit(value: 'grape', label: 'Grape', emoji: '🍇'),
-    Fruit(value: 'strawberry', label: 'Strawberry', emoji: '🍓'),
+    Fruit(value: 'apple', label: 'APPLE.EXE', emoji: '🍎'),
+    Fruit(value: 'banana', label: 'BANANA.SYS', emoji: '🍌'),
+    Fruit(value: 'orange', label: 'ORANGE.BAT', emoji: '🍊'),
+    Fruit(value: 'grape', label: 'GRAPE.DLL', emoji: '🍇'),
   ];
 
-  // Get selected fruit for display
-  Fruit? get _selectedFruit {
+  // Get selected fruit label for display
+  String? get _selectedLabel {
     if (_selectedValue == null) return null;
-    return fruits.firstWhere((f) => f.value == _selectedValue);
+    return fruits.firstWhere((f) => f.value == _selectedValue).label;
   }
 
-  Widget _buildOptionWithCheckmark(Fruit fruit) {
+  Widget _buildOption(Fruit fruit) {
     return NakedSelect.Option(
       value: fruit.value,
       builder: (context, state, _) {
-        final backgroundColor = state.when<Color?>(
-          selected: Colors.blue.shade50,
-          hovered: Colors.grey.shade100,
+        final backdropColor = state.when<Color>(
+          selected: const Color(0xFF001100),
+          hovered: const Color(0xFF001A00),
           orElse: Colors.transparent,
         );
-
+        final borderColor = state.when<Color>(
+          selected: const Color(0xFF00FF41),
+          hovered: const Color(0xFF00FF41),
+          orElse: Colors.transparent,
+        );
         final textColor = state.when<Color>(
-          selected: Colors.blue,
-          orElse: Colors.black,
+          selected: const Color(0xFF00FF41),
+          hovered: const Color(0xFF00AA33),
+          orElse: const Color(0xFF00AA33),
+        );
+        final isSelected = state.isSelected;
+        final isHovered = state.isHovered;
+
+        final textStyle = TextStyle(
+          color: textColor,
+          fontWeight: isSelected ? FontWeight.w600 : FontWeight.normal,
+          fontFamily: 'monospace',
         );
 
-        return AnimatedContainer(
-          duration: const Duration(milliseconds: 200),
-          curve: Curves.easeOut,
-          margin: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
-          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 12),
-          decoration: BoxDecoration(
-            color: backgroundColor,
-            borderRadius: BorderRadius.circular(8),
-          ),
-          child: Row(
-            children: [
-              Text(
-                fruit.emoji,
-                style: const TextStyle(fontSize: 20),
-              ),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Text(
-                  fruit.label,
-                  style: TextStyle(
-                    color: textColor,
-                    fontWeight:
-                        state.isSelected ? FontWeight.w600 : FontWeight.w500,
-                    fontSize: 14,
+        return Transform(
+          transform: Matrix4.identity(),
+          alignment: Alignment.centerLeft,
+          child: Container(
+            padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+            decoration: BoxDecoration(
+              color: backdropColor,
+              border: Border.all(color: borderColor, width: 1),
+              boxShadow: (isSelected || isHovered)
+                  ? [
+                      BoxShadow(
+                        color: const Color(0xFF00FF41).withValues(alpha: 0.3),
+                        blurRadius: 8,
+                        spreadRadius: 1,
+                      ),
+                    ]
+                  : null,
+            ),
+            child: Row(
+              spacing: 8,
+              children: [
+                ColorFiltered(
+                  colorFilter: const ColorFilter.mode(
+                    Color(0xFF00FF41),
+                    BlendMode.modulate,
+                  ),
+                  child: Text(
+                    fruit.emoji,
+                    style: TextStyle(
+                      fontSize: 16,
+                      shadows: isSelected || isHovered
+                          ? [
+                              const Shadow(
+                                color: Color(0xFF00FF41),
+                                blurRadius: 4,
+                              ),
+                            ]
+                          : null,
+                    ),
                   ),
                 ),
-              ),
-              AnimatedScale(
-                scale: state.isSelected ? 1.0 : 0.0,
-                duration: const Duration(milliseconds: 200),
-                curve: Curves.easeOutBack,
-                child: Container(
-                  padding: const EdgeInsets.all(2),
-                  decoration: const BoxDecoration(
-                    color: Colors.blue,
-                    shape: BoxShape.circle,
-                  ),
-                  child: const Icon(
-                    Icons.check,
-                    size: 14,
-                    color: Colors.white,
+                Expanded(
+                  child: Text(
+                    fruit.label,
+                    style: textStyle,
                   ),
                 ),
-              ),
-            ],
+                if (state.isSelected)
+                  const Icon(
+                    Icons.keyboard_arrow_right,
+                    size: 16,
+                    color: Color(0xFF00FF41),
+                  ),
+              ],
+            ),
           ),
         );
       },
@@ -150,7 +175,7 @@ class _CheckmarkSelectExampleState extends State<CheckmarkSelectExample> {
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      width: 280,
+      width: 250,
       child: NakedSelect<String>(
         value: _selectedValue,
         onChanged: (value) => setState(() => _selectedValue = value),
@@ -158,73 +183,133 @@ class _CheckmarkSelectExampleState extends State<CheckmarkSelectExample> {
           final focused = state.isFocused;
           final hovered = state.isHovered;
 
-          return AnimatedContainer(
-            duration: const Duration(milliseconds: 200),
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-            decoration: BoxDecoration(
-              color: Colors.white,
-              borderRadius: BorderRadius.circular(10),
-              border: Border.all(
-                color: focused ? Colors.blue : Colors.grey.shade300,
-              ),
-              boxShadow: [
-                BoxShadow(
-                  color: hovered
-                      ? const Color(0x14000000)
-                      : const Color(0x0A000000),
-                  blurRadius: hovered ? 8 : 4,
-                  offset: Offset(0, hovered ? 2 : 1),
+          return Transform(
+            transform: Matrix4.identity()
+              ..setEntry(3, 2, 0.001)
+              ..rotateY(-0.1),
+            alignment: Alignment.centerLeft,
+            child: Container(
+              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
+              decoration: BoxDecoration(
+                color: const Color(0xFF001100),
+                borderRadius: const BorderRadius.only(
+                  topLeft: Radius.circular(8),
+                  topRight: Radius.circular(0),
+                  bottomLeft: Radius.circular(0),
+                  bottomRight: Radius.circular(8),
                 ),
-              ],
-            ),
-            child: Row(
-              children: [
-                if (_selectedFruit != null) ...[
-                  Text(
-                    _selectedFruit!.emoji,
-                    style: const TextStyle(fontSize: 18),
+                border: Border.all(
+                  color: focused || hovered
+                      ? const Color(0xFF00FF41)
+                      : const Color(0xFF003300),
+                  width: 2,
+                ),
+                boxShadow: [
+                  if (focused || hovered)
+                    BoxShadow(
+                      color: const Color(0xFF00FF41)
+                          .withValues(alpha: focused ? 0.4 : 0.2),
+                      blurRadius: focused ? 12 : 8,
+                      spreadRadius: focused ? 2 : 1,
+                    ),
+                  BoxShadow(
+                    color: Colors.black.withValues(alpha: 0.3),
+                    blurRadius: 4,
+                    offset: const Offset(0, 2),
                   ),
-                  const SizedBox(width: 12),
                 ],
-                Expanded(
-                  child: Text(
-                    _selectedFruit?.label ?? 'Choose your favorite fruit...',
-                    style: TextStyle(
-                      color:
-                          _selectedFruit != null ? Colors.black : Colors.grey,
-                      fontSize: 14,
-                      fontWeight: FontWeight.w500,
+              ),
+              child: Row(
+                children: [
+                  if (_selectedValue != null) ...[
+                    ColorFiltered(
+                      colorFilter: const ColorFilter.mode(
+                        Color(0xFF00FF41),
+                        BlendMode.modulate,
+                      ),
+                      child: Text(
+                        fruits
+                            .firstWhere((f) => f.value == _selectedValue)
+                            .emoji,
+                        style: const TextStyle(
+                          fontSize: 16,
+                          shadows: [
+                            Shadow(
+                              color: Color(0xFF00FF41),
+                              blurRadius: 4,
+                            ),
+                          ],
+                        ),
+                      ),
+                    ),
+                    const SizedBox(width: 8),
+                  ],
+                  Expanded(
+                    child: Text(
+                      _selectedLabel ?? '> SELECT DATA_',
+                      style: TextStyle(
+                        color: _selectedValue != null
+                            ? const Color(0xFF00FF41)
+                            : const Color(0xFF00AA33),
+                        fontFamily: 'monospace',
+                        fontSize: 13,
+                        fontWeight: FontWeight.w500,
+                      ),
                     ),
                   ),
-                ),
-                const Icon(
-                  Icons.expand_more,
-                  size: 20,
-                  color: Colors.grey,
-                ),
-              ],
+                  Icon(
+                    Icons.keyboard_arrow_down,
+                    size: 20,
+                    color: focused
+                        ? const Color(0xFF00FF41)
+                        : const Color(0xFF00AA33),
+                  ),
+                ],
+              ),
             ),
           );
         },
         overlayBuilder: (context, info) {
-          return Container(
-            margin: const EdgeInsets.only(top: 4),
-            padding: const EdgeInsets.symmetric(vertical: 8),
-            decoration: BoxDecoration(
-              color: Colors.white,
-              borderRadius: BorderRadius.circular(10),
-              border: Border.all(color: const Color(0xFFE2E8F0)),
-              boxShadow: const [
-                BoxShadow(
-                  color: Color(0x14000000),
-                  blurRadius: 20,
-                  offset: Offset(0, 8),
+          return SizedBox(
+            width: 250,
+            child: Transform(
+              transform: Matrix4.identity()
+                ..setEntry(3, 2, 0.001)
+                ..rotateY(-0.1),
+              alignment: Alignment.centerLeft,
+              child: Container(
+                margin: const EdgeInsets.only(top: 4),
+                padding: const EdgeInsets.symmetric(vertical: 6),
+                decoration: BoxDecoration(
+                  color: const Color(0xFF000A00),
+                  borderRadius: const BorderRadius.only(
+                    topLeft: Radius.circular(8),
+                    topRight: Radius.circular(0),
+                    bottomLeft: Radius.circular(0),
+                    bottomRight: Radius.circular(8),
+                  ),
+                  border: Border.all(
+                    color: const Color(0xFF00FF41),
+                    width: 2,
+                  ),
+                  boxShadow: [
+                    BoxShadow(
+                      color: const Color(0xFF00FF41).withValues(alpha: 0.3),
+                      blurRadius: 16,
+                      spreadRadius: 2,
+                    ),
+                    BoxShadow(
+                      color: Colors.black.withValues(alpha: 0.5),
+                      blurRadius: 12,
+                      offset: const Offset(0, 4),
+                    ),
+                  ],
                 ),
-              ],
-            ),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: fruits.map(_buildOptionWithCheckmark).toList(),
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: fruits.map(_buildOption).toList(),
+                ),
+              ),
             ),
           );
         },
