+++ example/lib/api/naked_toggle.0.dart
@@ -1,6 +1,47 @@
 import 'package:flutter/material.dart';
 import 'package:naked_ui/naked_ui.dart';
 
+/// Main function
+void main() {
+  runApp(const MyApp());
+}
+
+/// Main App
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        backgroundColor: Colors.grey.shade50,
+        body: const Center(
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text(
+                'Toggle Button Example',
+                style: TextStyle(
+                  fontSize: 24,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+              SizedBox(height: 8),
+              Text(
+                'Interact with the toggle button to see its states',
+                style: TextStyle(color: Colors.grey),
+              ),
+              SizedBox(height: 24),
+              ToggleButtonExample(),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+/// Toggle Button Example
 class ToggleButtonExample extends StatefulWidget {
   const ToggleButtonExample({super.key});
 
@@ -8,6 +49,7 @@ class ToggleButtonExample extends StatefulWidget {
   State<ToggleButtonExample> createState() => _ToggleButtonExampleState();
 }
 
+/// Toggle Button Example State
 class _ToggleButtonExampleState extends State<ToggleButtonExample> {
   bool _isBold = false;
   bool _isItalic = false;
@@ -36,13 +78,13 @@ class _ToggleButtonExampleState extends State<ToggleButtonExample> {
           height: 40,
           decoration: BoxDecoration(
             color: isSelected
-                ? Colors.blue.shade600
+                ? Colors.grey.shade800
                 : isHovered
                     ? Colors.grey.shade200
-                    : Colors.transparent,
+                    : Colors.white,
             borderRadius: BorderRadius.circular(6),
             border: isFocused
-                ? Border.all(color: Colors.blue, width: 2)
+                ? Border.all(color: Colors.grey.shade400, width: 2)
                 : Border.all(color: Colors.grey.shade300),
             boxShadow: isPressed
                 ? [
@@ -66,85 +108,31 @@ class _ToggleButtonExampleState extends State<ToggleButtonExample> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Toggle Button Example'),
-      ),
-      body: Center(
-        child: Container(
-          padding: const EdgeInsets.all(24),
-          margin: const EdgeInsets.all(16),
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.circular(12),
-            border: Border.all(color: Colors.grey.shade300),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black.withValues(alpha: 0.1),
-                blurRadius: 8,
-                offset: const Offset(0, 2),
-              ),
-            ],
+    return Center(
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          _buildToggleButton(
+            icon: Icons.format_bold,
+            isSelected: _isBold,
+            onChanged: (value) => setState(() => _isBold = value),
+            tooltip: 'Bold',
           ),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              const Text(
-                'Text Formatting',
-                style: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              const SizedBox(height: 16),
-              Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  _buildToggleButton(
-                    icon: Icons.format_bold,
-                    isSelected: _isBold,
-                    onChanged: (value) => setState(() => _isBold = value),
-                    tooltip: 'Bold',
-                  ),
-                  const SizedBox(width: 8),
-                  _buildToggleButton(
-                    icon: Icons.format_italic,
-                    isSelected: _isItalic,
-                    onChanged: (value) => setState(() => _isItalic = value),
-                    tooltip: 'Italic',
-                  ),
-                  const SizedBox(width: 8),
-                  _buildToggleButton(
-                    icon: Icons.format_underlined,
-                    isSelected: _isUnderlined,
-                    onChanged: (value) => setState(() => _isUnderlined = value),
-                    tooltip: 'Underline',
-                  ),
-                ],
-              ),
-              const SizedBox(height: 20),
-              Container(
-                width: double.infinity,
-                padding: const EdgeInsets.all(16),
-                decoration: BoxDecoration(
-                  color: Colors.grey.shade50,
-                  borderRadius: BorderRadius.circular(8),
-                  border: Border.all(color: Colors.grey.shade200),
-                ),
-                child: Text(
-                  'Sample Text',
-                  style: TextStyle(
-                    fontSize: 16,
-                    fontWeight: _isBold ? FontWeight.bold : FontWeight.normal,
-                    fontStyle: _isItalic ? FontStyle.italic : FontStyle.normal,
-                    decoration: _isUnderlined ? TextDecoration.underline : null,
-                  ),
-                ),
-              ),
-            ],
+          const SizedBox(width: 8),
+          _buildToggleButton(
+            icon: Icons.format_italic,
+            isSelected: _isItalic,
+            onChanged: (value) => setState(() => _isItalic = value),
+            tooltip: 'Italic',
           ),
-        ),
+          const SizedBox(width: 8),
+          _buildToggleButton(
+            icon: Icons.format_underlined,
+            isSelected: _isUnderlined,
+            onChanged: (value) => setState(() => _isUnderlined = value),
+            tooltip: 'Underline',
+          ),
+        ],
       ),
     );
   }
